+++ src/physics/managers/multi/src/phy_modules.cu
@@ -6,6 +6,7 @@
 
 #include "phy_modules.h"
 
+#include "boundary_layer.h"
 #include "log_writer.h"
 #include "radiative_transfer.h"
 
@@ -20,18 +21,29 @@ const bool radiative_transfer_enabled_default = false;
 
 radiative_transfer rt;
 
+bool       boundary_layer_enabled         = false;
+const bool boundary_layer_enabled_default = false;
+
+boundary_layer bl;
+
 
 std::string phy_modules_get_name() {
     return std::string("multi");
 }
 
 void phy_modules_print_config() {
-    log::printf("  multi physics module, with radiative transfer \n");
+    log::printf("  multi physics module, with radiative transfer, boundary_layer \n");
     log::printf("   Radiative Transfer module: %s.\n",
                 radiative_transfer_enabled ? "true" : "false");
 
+    log::printf("   Boundary Layer module: %s.\n", boundary_layer_enabled ? "true" : "false");
+    log::printf("\n");
+
     if (radiative_transfer_enabled)
         rt.print_config();
+
+    if (boundary_layer_enabled)
+        bl.print_config();
 }
 
 
@@ -42,6 +54,9 @@ bool phy_modules_init_mem(const ESP& esp, device_RK_array_manager& phy_modules_c
 
     if (radiative_transfer_enabled)
         rt.initialise_memory(esp, phy_modules_core_arrays);
+
+    if (boundary_layer_enabled)
+        bl.initialise_memory(esp, phy_modules_core_arrays);
 
     return out;
 }
@@ -57,6 +72,9 @@ bool phy_modules_init_data(const ESP& esp, const SimulationSetup& sim, storage* 
     if (radiative_transfer_enabled)
         out &= rt.initial_conditions(esp, sim);
 
+    if (boundary_layer_enabled)
+        out &= bl.initial_conditions(esp, sim);
+
     return out;
 }
 
@@ -67,6 +85,11 @@ bool phy_modules_generate_config(config_file& config_reader) {
         "radiative_transfer", radiative_transfer_enabled, radiative_transfer_enabled_default);
 
     rt.configure(config_reader);
+
+    config_reader.append_config_var(
+        "boundary_layer", boundary_layer_enabled, boundary_layer_enabled_default);
+
+    bl.configure(config_reader);
 
     return out;
 }
@@ -105,17 +128,25 @@ bool phy_modules_phy_loop(ESP& esp, const SimulationSetup& sim, int nstep, doubl
     if (radiative_transfer_enabled)
         rt.phy_loop(esp, sim, nstep, time_step);
 
+    if (boundary_layer_enabled)
+        bl.phy_loop(esp, sim, nstep, time_step);
+
     return out;
 }
 
 bool phy_modules_store_init(storage& s) {
     // radiative transfer option
     s.append_value(radiative_transfer_enabled ? 1.0 : 0.0,
-                   "/radiativetransfer",
+                   "/radiative_transfer",
                    "-",
                    "Using radiative transfer");
 
     rt.store_init(s);
+
+    s.append_value(
+        boundary_layer_enabled ? 1.0 : 0.0, "/boundary_layer", "-", "Using boundary layer");
+
+    bl.store_init(s);
 
     return true;
 }
@@ -124,6 +155,9 @@ bool phy_modules_store(const ESP& esp, storage& s) {
 
     if (radiative_transfer_enabled)
         rt.store(esp, s);
+
+    if (boundary_layer_enabled)
+        bl.store(esp, s);
 
     return true;
 }
@@ -135,6 +169,9 @@ bool phy_modules_free_mem() {
 
     if (radiative_transfer_enabled)
         rt.free_memory();
+
+    if (boundary_layer_enabled)
+        bl.free_memory();
 
     return out;
 }