+++ src/physics/modules/inc/radiative_transfer.h
@@ -75,12 +75,17 @@ private:
     double Tstar_config            = 4520;  // Star effective temperature [K]
     double planet_star_dist_config = 0.015; // Planet-star distance [au]
     double radius_star_config      = 0.667; // Star radius [Rsun]
-    double diff_fac_config         = 0.5;   // Diffusivity factor: 0.5-1.0
+    double diff_ang_config         = 0.5;   // Diffusivity angle (1 / diffusivity factor): 0.5-1.0
     double Tlow_config =
         970; // Lower boundary temperature: upward flux coming from the planet's interior
-    double albedo_config = 0.18;   // Bond albedo
-    double tausw_config  = 532.0;  // Absorption coefficient for the shortwaves
-    double taulw_config  = 1064.0; // Absorption coefficient for the longwaves
+    double albedo_config     = 0.18;   // Bond albedo
+    double tausw_config      = 532.0;  // Absorption coefficient for the shortwaves
+    double taulw_config      = 1064.0; // Absorption coefficient for the longwaves
+    bool   latf_lw_config    = false;  // use sin^2(lat) dependence for lw opacity
+    double taulw_pole_config = 1064.0; // Absorption coefficient for the longwave (poles)
+    double n_lw_config       = 2.0;    // power law dependence for unmixed absorbers in LW
+    double n_sw_config       = 1.0;    // power law dependence for mixed/unmixed absorbers in SW
+    double f_lw_config       = 0.5;    // fraction of taulw in well-mixed absorber
 
 
     bool   sync_rot_config    = true;     // is planet syncronously rotating?
@@ -91,16 +96,30 @@ private:
     double alpha_i_config     = 0;        // initial right asc of host star (relative to long = 0)
     double longp_config       = 0;        // longitude of periastron (rad)
 
+    bool   surface_config = false; // use solid/liquid surface at altitude 0
+    double Csurf_config   = 1e7;   // heat capacity of surface (J K^-1 m^-2)
 
     // Rad trans
     double Tstar            = 4520;  // Star effective temperature [K]
     double planet_star_dist = 0.015; // Planet-star distance [au]
     double radius_star      = 0.667; // Star radius [Rsun]
-    double diff_fac         = 0.5;   // Diffusivity factor: 0.5-1.0
+    double diff_ang         = 0.5;   // Diffusivity factor: 0.5-1.0
     double Tlow = 970; // Lower boundary temperature: upward flux coming from the planet's interior
-    double albedo = 0.18;   // Bond albedo
-    double tausw  = 532.0;  // Absorption coefficient for the shortwaves
-    double taulw  = 1064.0; // Absorption coefficient for the longwaves
+    double albedo     = 0.18;   // Bond albedo
+    double tausw      = 532.0;  // Absorption coefficient for the shortwaves
+    double taulw      = 1064.0; // Absorption coefficient for the longwaves
+    double taulw_pole = 1064.0;
+    bool   latf_lw    = false;
+    double n_lw       = 2.0; // power law dependence for unmixed absorbers in LW
+    double n_sw       = 1.0; // power law dependence for mixed/unmixed absorbers in SW
+    double f_lw       = 0.5; // fraction of taulw in well-mixed absorber
+
+    bool    surface;
+    double  Csurf;
+    double *surf_flux_d;
+    double *Tsurface_d;
+    double *Tsurface_h;
+
     double incflx;
     //  Arrays used in RT code
     double *fnet_up_d;
@@ -140,11 +159,16 @@ private:
     void    RTSetup(double Tstar_,
                     double planet_star_dist_,
                     double radius_star_,
-                    double diff_fac_,
+                    double diff_ang_,
                     double Tlow_,
                     double albedo_,
                     double tausw_,
                     double taulw_,
+                    bool   latf_lw_,
+                    double taulw_pole_,
+                    double n_lw_,
+                    double n_sw_,
+                    double f_lw_,
                     bool   sync_rot_,
                     double mean_motion_,
                     double true_long_i_,
@@ -153,6 +177,9 @@ private:
                     double alpha_i_,
                     double obliquity_,
                     double Omega,
+                    bool   surface,
+                    double Csurf,
+                    double Tmean,
                     int    point_num);
 
     void update_spin_orbit(double time, double Omega);